Refactored question types into separate files with signalR included, fixed minor errors.
