+++ backend/TriviaBackend/Controllers/GameController.cs
@@ -1,6 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using TriviaBackend.Models;
 using TriviaBackend.Services;
+using TriviaBackend.Models.Enums;
+using TriviaBackend.Models.Records;
+using TriviaBackend.Models.Objects;
 
 namespace TriviaBackend.Controllers
 {
@@ -12,7 +15,7 @@ namespace TriviaBackend.Controllers
 
         public GameController()
         {
-            _settings = new GameSettings(maxPlayers: 5, questionsPerGame: 10, defaultTimeLimit: 20);
+            _settings = new GameSettings(MaxPlayers: 5, QuestionsPerGame: 10, DefaultTimeLimit: 20);
             _questionService = new QuestionService();
             _gameEngine = new GameEngine(_questionService, _settings);
         }
