+++ backend/TriviaBackend/Models/Enums/AnswerResult.cs
@@ -1,0 +1,9 @@
+namespace TriviaBackend.Models.Enums
+{
+    public enum AnswerResult
+    {
+        Correct,
+        Incorrect,
+        TimeUp
+    }
+}
