+++ backend/TriviaBackend/Models/Enums/DifficultyLevel.cs
@@ -1,0 +1,9 @@
+namespace TriviaBackend.Models.Enums
+{
+    public enum DifficultyLevel
+    {
+        Easy = 1,
+        Medium = 2,
+        Hard = 3
+    }
+}
