+++ backend/TriviaBackend/Models/Enums/GameStatus.cs
@@ -1,0 +1,10 @@
+namespace TriviaBackend.Models.Enums
+{
+    public enum GameStatus
+    {
+        Waiting,
+        InProgress,
+        Finished,
+        Paused
+    }
+}
