+++ backend/TriviaBackend/Models/Enums/QuestionCategory.cs
@@ -1,0 +1,11 @@
+namespace TriviaBackend.Models.Enums
+{
+    public enum QuestionCategory
+    {
+        Science,
+        History,
+        Sports,
+        Geography,
+        Literature
+    }
+}
