--- backend/TriviaBackend/Models/Question.cs
@@ -1,66 +1,0 @@
-﻿namespace TriviaBackend.Models
-{
-    public enum QuestionCategory
-    {
-        Science,
-        History,
-        Sports,
-        Geography,
-        Literature
-    }
-    public enum DifficultyLevel
-    {
-        Easy = 1,
-        Medium = 2,
-        Hard = 3
-    }
-    public enum GameStatus
-    {
-        Waiting,
-        InProgress,
-        Finished,
-        Paused
-    }
-    public enum AnswerStatus
-    {
-        Correct,
-        Incorrect,
-        Timesup
-    }
-    public class TriviaQuestion
-    {
-        public int Id { get; set; }
-        public string QuestionText { get; set; }
-        public string Text => QuestionText;  // fix: signalR invoked, but property not found
-        public List<string> Options { get; set; }
-        public int CorrectAnswerIndex { get; set; }
-        public QuestionCategory Category { get; set; }
-        public DifficultyLevel Difficulty { get; set; }
-        public int TimeLimit { get; set; } = 30;
-        public int Points => (int)Difficulty * 10;
-
-        public TriviaQuestion()
-        {
-            Options = new List<string>();
-        }
-    }
-    public record PlayerAnswer(int QuestionId, int SelectedOptionIndex, AnswerStatus Status, int PointsEarned);
-
-    public struct GameSettings
-    {
-        public int MaxPlayers { get; set; }
-        public int QuestionsPerGame { get; set; }
-        public int DefaultTimeLimit { get; set; }
-        public bool AllowLateJoining { get; set; }
-        public QuestionCategory[] questionCategories { get; set; }
-
-        public GameSettings(int maxPlayers = 10, int questionsPerGame = 10, int defaultTimeLimit = 30)
-        {
-            this.MaxPlayers = maxPlayers;
-            this.QuestionsPerGame = questionsPerGame;
-            this.DefaultTimeLimit = defaultTimeLimit;
-            this.AllowLateJoining = false;
-            this.questionCategories = Enum.GetValues<QuestionCategory>();
-        }
-    }
-}
