+++ backend/TriviaBackend/Models/Records/GameAnswer.cs
@@ -1,0 +1,4 @@
+namespace TriviaBackend.Models.Records
+{
+    public record GameAnswer(int PlayerId, int QuestionId, int SelectedOptionIndex, DateTime SubmissionTime);
+}
