+++ backend/TriviaBackend/Models/Records/PlayerAnswer.cs
@@ -1,0 +1,6 @@
+using TriviaBackend.Models.Enums;
+
+namespace TriviaBackend.Models.Records
+{
+    public record PlayerAnswer(int QuestionId, int SelectedOptionIndex, AnswerResult Status, int PointsEarned);
+}
