+++ backend/TriviaBackend/Services/GameEngineService.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using TriviaBackend.Models;
+using TriviaBackend.Models.Enums;
+using TriviaBackend.Models.Records;
+using TriviaBackend.Models.Objects;
 
 namespace TriviaBackend.Services
 {
